Refactor Navbar components for improved structure and functionality; consolidate NavContent into NavList, enhance menu toggle behavior, and update styles for better responsiveness.
