+++ src/assets/scss/components/Navbar.scss
@@ -18,11 +18,8 @@
     translate: 0 -100%;
   }
 
-  &__content {
+  .container {
     @include breakpoint-up(md) {
-      position: relative;
-      display: flex;
-
       &::before {
         content: "";
         position: absolute;
@@ -36,49 +33,55 @@
     }
   }
 
+  &__list {
+    padding: 1.5rem 0;
+
+    &--item {
+      border-radius: var(--radius-200);
+
+      &__link {
+        padding: 0.5rem;
+        display: inline-block;
+        color: var(--text-400);
+        font-size: clamp(0.75em, 2.5vw, 1em);
+        font-weight: 500;
+        letter-spacing: 1.5px;
+        transition: color 100ms ease-in-out;
+
+        &:hover {
+          color: var(--text-100);
+        }
+      }
+    }
+  }
+
   &__menu {
+    position: absolute;
+    top: 6rem;
+    right: 1rem;
+    left: 1rem;
+
     text-align: center;
 
-    @include breakpoint-up(md) {
-      display: flex;
-      padding: 1.5rem 0;
-    }
-
     &-toggle {
-      flex-direction: column;
-      gap: 0.25rem;
+      gap: 0.3rem;
+      padding: 0.35rem;
       border-radius: var(--radius-100);
-      background-color: transparent;
+      background-color: var((--opaque-light-200));
 
       &--stroke {
         position: relative;
         display: inline-block;
-        height: 0.45rem;
-        width: 2.75rem;
-        background: var(--primary-gradient);
+        height: 0.2rem;
+        width: 1.65rem;
+        background-color: var(--text-300);
         border-radius: var(--radius-pill);
-        border: solid 2px var(--text-100);
-
-        &::before {
-          content: "";
-          position: absolute;
-          inset: 0;
-          background-color: var(--bg-100);
-          border-radius: inherit;
-          opacity: 1;
-          transition: opacity 150ms ease-in-out;
-        }
-      }
-
-      @include breakpoint-up(large) {
-        display: none;
+        transition: background-color 150ms ease-in-out;
       }
 
       &:hover {
         .nav__menu-toggle--stroke {
-          &::before {
-            opacity: 0;
-          }
+          background-color: var(--text-100);
         }
       }
     }
@@ -111,7 +114,6 @@
       }
 
       &__content {
-        position: relative;
         font-size: 1em;
       }
 
@@ -133,34 +135,28 @@
       &__popup {
         position: absolute;
         right: 0;
-        top: 105%;
+        top: 2.75rem;
         width: max-content;
         border-radius: var(--radius-200);
-        background-color: var(--opaque-300);
-        border: solid 0.125em var(--opaque-200);
+        background-color: var(--bg-300);
+        border: solid 0.125em var(--opaque-20);
         z-index: 1;
       }
     }
 
-    li {
+    &--item {
       position: relative;
       border-radius: var(--radius-200);
       background-color: var(--bg-200);
 
-      @include breakpoint-up(md) {
-        background-color: transparent;
-      }
-
       &:not(:first-child) {
         margin-top: 0.25rem;
       }
-    }
 
-    &-item {
-      a {
+      &__link {
         display: block;
         color: var(--text-400);
-        cursor: pointer;
+        font-size: 1.25rem;
         font-weight: 500;
         letter-spacing: 1.5px;
         transition: color 100ms ease-in-out, background-color 100ms ease-in-out;
@@ -175,25 +171,6 @@
           background-color: var(--bg-400);
         }
 
-        @include breakpoint-down(sm) {
-          font-size: 1.25rem;
-        }
-
-        @include breakpoint-up(md) {
-          padding: 0.5rem;
-          font-size: clamp(0.75em, 2.5vw, 1em);
-
-          &:hover {
-            color: var(--text-100);
-            background-color: transparent;
-          }
-
-          &:active {
-            color: var(--text-100);
-            background-color: transparent;
-          }
-        }
-
         &.active {
           color: var(--text-100);
         }
@@ -206,12 +183,11 @@
       &__menu {
         &-toggle {
           &--stroke {
-            border: solid 2px var(--text-200);
             transition: opacity 0ms 100ms ease-in-out, top 100ms ease-in-out,
               rotate 100ms 200ms ease-in-out;
 
             &-top {
-              top: 11px;
+              top: 8px;
               rotate: 45deg;
             }
 
@@ -220,7 +196,7 @@
             }
 
             &-bottom {
-              top: -11px;
+              top: -8px;
               rotate: -45deg;
             }
           }
