+++ src/components/Navbar/MenuToggleButton.jsx
@@ -5,11 +5,11 @@ const MenuToggleButton = () => {
   const { toggleMenu } = useContext(MenuContext);
 
   return (
-    <div className="nav__menu-toggle--wrap p-1">
+    <div className="nav__menu-toggle--wrap py-3">
       <button
         type="button"
         onClick={toggleMenu}
-        className="nav__menu-toggle flex fle-jc-c flex-ai-c"
+        className="nav__menu-toggle flex flex-col fle-jc-c flex-ai-c"
       >
         <span className="nav__menu-toggle--stroke nav__menu-toggle--stroke-top" />
         <span className="nav__menu-toggle--stroke nav__menu-toggle--stroke-mid" />
